requirements.txt
asgiref==3.8.1
bleach==6.1.0
chardet==5.2.0
coverage==7.6.1
diff-match-patch==20230430
Django==4.2
django-cors-headers==4.4.0
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.1
inflection==0.5.1
mccabe==0.7.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tinycss2==1.3.0
tzdata==2024.1
uritemplate==4.1.1
webencodings==0.5.1
